extract PasswordStrengthMeter to standalone component
