+++ portal/src/PasswordField.module.scss
@@ -1,11 +1,3 @@
-$passwordMeterInactiveColor: #e0e0e0;
-$passwordMeterLevel1Color: #e23d3d;
-$passwordMeterLevel2Color: #e67c00;
-$passwordMeterLevel3Color: #d69900;
-$passwordMeterLevel4Color: #93a03f;
-$passwordMeterLevel5Color: #04ab18;
-$passwordPolicySatisfiedColor: #04ab18;
-
 .passwordPolicy {
   margin-top: 10px;
 }
@@ -15,138 +7,5 @@ $passwordPolicySatisfiedColor: #04ab18;
 }
 
 .policySatisfied {
-  color: $passwordPolicySatisfiedColor;
-}
-
-.passwordStrengthMeter {
-  appearance: none;
-  display: block;
-  width: 266px;
-  height: 4px;
-  margin: 15px 0 8px 2px;
-}
-
-.passwordStrengthMeterDescriptionContainer {
-  margin: 0 0 15px 2px;
-}
-.passwordStrengthMeterDescriptionLevel1 {
-  color: $passwordMeterLevel1Color;
-}
-
-.passwordStrengthMeterDescriptionLevel2 {
-  color: $passwordMeterLevel2Color;
-}
-
-.passwordStrengthMeterDescriptionLevel3 {
-  color: $passwordMeterLevel3Color;
-}
-
-.passwordStrengthMeterDescriptionLevel4 {
-  color: $passwordMeterLevel4Color;
-}
-
-.passwordStrengthMeterDescriptionLevel5 {
-  color: $passwordMeterLevel5Color;
-}
-
-.passwordStrengthMeter,
-.passwordStrengthMeter::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterInactiveColor 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterInactiveColor 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterInactiveColor 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterInactiveColor 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterInactiveColor 81.6% 100%
-  );
-}
-
-.passwordStrengthMeter::-webkit-meter-optimum-value,
-.passwordStrengthMeter::-webkit-meter-suboptimum-value,
-.passwordStrengthMeter::-webkit-meter-even-less-good-value {
-  background: none;
-}
-
-.passwordStrengthMeter[value="1"],
-.passwordStrengthMeter[value="1"]::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterLevel1Color 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterInactiveColor 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterInactiveColor 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterInactiveColor 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterInactiveColor 81.6% 100%
-  );
-}
-
-.passwordStrengthMeter[value="2"],
-.passwordStrengthMeter[value="2"]::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterLevel2Color 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterLevel2Color 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterInactiveColor 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterInactiveColor 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterInactiveColor 81.6% 100%
-  );
-}
-
-.passwordStrengthMeter[value="3"],
-.passwordStrengthMeter[value="3"]::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterLevel3Color 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterLevel3Color 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterLevel3Color 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterInactiveColor 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterInactiveColor 81.6% 100%
-  );
-}
-
-.passwordStrengthMeter[value="4"],
-.passwordStrengthMeter[value="4"]::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterLevel4Color 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterLevel4Color 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterLevel4Color 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterLevel4Color 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterInactiveColor 81.6% 100%
-  );
-}
-
-.passwordStrengthMeter[value="5"],
-.passwordStrengthMeter[value="5"]::-webkit-meter-bar {
-  background: linear-gradient(
-    90deg,
-    $passwordMeterLevel5Color 0% 18.4%,
-    transparent 18.4% 20.4%,
-    $passwordMeterLevel5Color 20.4% 38.8%,
-    transparent 38.8% 40.8%,
-    $passwordMeterLevel5Color 40.8% 59.2%,
-    transparent 59.2% 61.2%,
-    $passwordMeterLevel5Color 61.2% 79.6%,
-    transparent 79.6% 81.6%,
-    $passwordMeterLevel5Color 81.6% 100%
-  );
+  color: #04ab18;
 }
